Create function generateMajorScale which stores the note for each scale
Having the notes for each scale stored in an array, sorted by their order in the scale,
will allow chords to be generated based on their numerical values in the scale. For instance, the root major chord would be the 1, 3, and 5 notes. The IV chord would be the 4, 6, and 1 notes.