Create Chord objects from within Scale object
Chord objects changed to *not* inherit from Monobehavior, so they can be init'd with 'new' keyword from within Scale.cs, and without having to each be attached to a GameObject. ChordNames array parrallels the name property in Chord, but allows display in Inspector. Scale is associated with just one game object, which can update to a different scale as needed.