Update Fretboard from the Scale object
When the scale is changed through the scale dropdown, the Fretboard object calls the Scales object to get info about the scale, and then changes the active status of each of the notes so that only the notes of the current scale are active. The root chord of the scale is selected by default, setting only the notes of that chord to colored, and the rest to gray.