Add spacing and offset properties to GuitarString
The spacing is the distance between frets. The offset is the amount to shift the position of the notes left so that they are between the frets. Coresponding public properties are in Fretboard, where they are set. The values in Fretboard are passed to each GuitarString when they are init'd.