Enable Scales and Chords with the Notes in GuitarString
Disabled changing scale and chords with the previous array of Notes attached directly to fretboard. Enabled changing the scale and chords with the array of Notes attached to each GuitarString's.
To follow the music nomenclature and guitar tablature commonly used, arranged the strings so they lay out in the negative y direction, so that the first string is towards the top of the screen, and the 6th string is towards the bottom of the screen(corresponding to a physical guitar the opposite, with the 1st string pointed towards the floor).