Draw ChordShape based on position of Note objects
Get the fret number which contains the root note of the chord on the 1st string. Contruct the G-shape chord from that note, the corresponding note on the 6th string, and one fret below it on the 5th string. Create a new Chord object with those coordinates. Draw a chord shape for it, using the VectorChord script's DrawChord function.