Refactor GetChordShapes to cover 3 octaves
Refactor GetChordShapes in ChordShapes, so that it covers the entire fretboard, looping over 3 octaves, so the chord shapes are displayed for all 3 octaves. However, if the note for a point is out of the range of the fretboard, it is not added to the chord shape path. In the case where one or more notes are out of range, the chord path is checked for empty points, which are removed (eliminating problems such as a chord that is cut off at the upper end draws its final point back to the 1st fret).