Align strings and notes to guitar neck model
The strings and their notes are now positioned according to the new guitar neck model, rather than the previous placeholder background prefab array. The background prefabs are no longer being generated, replaced by a model of an actual guitar. The notes are now aligned with the fret positions (the frets and their notes getting closer together the further you go up the fretboard).