Leave hidden notes active, but disable renderer
Notes that are not in the current scale, and therefore are hidden, are now playable. Instead of setting setActive to false, which hides the note and inactivates the audio source as well, the renderer for both the Note and its textmesh are disabled. All notes on the fretboard can now be played, even when they are not in the current scale. Relevant code is in SetActiveNotesForScale in Fretboard.cs.