Fix UI scaling and positioning
Move the canvas element that holds the chord vector shapes up in Y by 1, so that the chord shapes don't disappear behind the guitar object.
Change the canvas UI scaling to scale with screen size, and reduce the size of the UI components, so that the menus are a reasonable size on both iPhone 6s Plus, and iPhone 5s. Remove the background frame for the UI, as all available space is needed at top so that the notes can be reached by fingers.