Fix note playback latency lag
A simple fix: In the Audio manager under project settings, change the DSP buffer to 'Best Latency' ('Good Latency' might also work.) This makes the DSP Audio buffer smaller, which can result in less smooth playback, but improves latency. When playing notes on the fretboard, it now has the response of a real guitar, no lag. There is sometimes a decrease in quality, so trying to adjust the buffer settings may be worthwhile to find a good balance.